refactor film and follower features to use MVI architecture and improve module organization
