+++ commonApp/src/commonMain/kotlin/com/stslex/wizard/di/AppModule.kt
@@ -6,7 +6,6 @@ import com.stslex.wizard.core.database.di.ModuleCoreDatabase
 import com.stslex.wizard.core.network.api.di.ModuleCoreNetworkApi
 import com.stslex.wizard.core.network.client.di.ModuleCoreNetworkClient
 import com.stslex.wizard.core.ui.kit.di.ModuleCoreUi
-import com.stslex.wizard.feature.film.di.ModuleFeatureFilm
 import com.stslex.wizard.feature.follower.di.ModuleFeatureFollower
 import com.stslex.wizard.feature.match.di.ModuleFeatureMatch
 import com.stslex.wizard.feature.match_feed.di.ModuleFeatureMatchFeed
@@ -19,7 +18,6 @@ val appModules = listOf(
     ModuleCoreNetworkApi(),
     ModuleCoreDatabase(),
     ModuleFeatureMatchFeed(),
-    ModuleFeatureFilm(),
     ModuleFeatureFollower(),
     ModuleFeatureSettings(),
     ModuleFeatureMatch()
