+++ feature/film_feed/src/commonMain/kotlin/com/stslex/wizard/feature/film_feed/di/ModuleFeatureFilmFeed.kt
@@ -6,12 +6,4 @@ import org.koin.core.annotation.Module
 @Module
 @ComponentScan("com.stslex.wizard.feature.film_feed")
 class ModuleFeatureFilmFeed
-//    : AppModule() {
-//
-//    override fun declaration(): ModuleDeclaration = {
-//        viewModelOf(::FeedStoreImpl) { bind<FeedStore>() }
-//        factoryOf(::FeedInteractorImpl) { bind<FeedInteractor>() }
-//        factoryOf(::FeedRepositoryImpl) { bind<FeedRepository>() }
-//    }
-//}
