+++ feature/film_feed/src/commonMain/kotlin/com/stslex/wizard/feature/film_feed/mvi/ClickersHandler.kt
@@ -2,9 +2,8 @@ package com.stslex.wizard.feature.film_feed.mvi
 
 import com.stslex.wizard.core.ui.mvi.v2.Handler
 import com.stslex.wizard.feature.film_feed.di.FilmFeedScope
-import com.stslex.wizard.feature.film_feed.ui.store.FeedHandlerStore
-import com.stslex.wizard.feature.film_feed.ui.store.FeedStore.Action
-import com.stslex.wizard.feature.film_feed.ui.store.FeedStore.Action.Click
+import com.stslex.wizard.feature.film_feed.mvi.FeedStore.Action
+import com.stslex.wizard.feature.film_feed.mvi.FeedStore.Action.Click
 import org.koin.core.annotation.Factory
 import org.koin.core.annotation.Scope
 import org.koin.core.annotation.Scoped
