+++ feature/film_feed/src/commonMain/kotlin/com/stslex/wizard/feature/film_feed/mvi/FeedHandlerStore.kt
@@ -1,0 +1,5 @@
+package com.stslex.wizard.feature.film_feed.mvi
+
+import com.stslex.wizard.core.ui.mvi.v2.HandlerStore
+
+interface FeedHandlerStore : FeedStore, HandlerStore<FeedStore.State, FeedStore.Action, FeedStore.Event>
