+++ feature/film_feed/src/commonMain/kotlin/com/stslex/wizard/feature/film_feed/mvi/FeedStore.kt
@@ -1,4 +1,4 @@
-package com.stslex.wizard.feature.film_feed.ui.store
+package com.stslex.wizard.feature.film_feed.mvi
 
 import androidx.compose.runtime.Stable
 import com.stslex.wizard.core.navigation.Target
@@ -7,9 +7,9 @@ import com.stslex.wizard.core.ui.kit.pager.PagingUIList.Companion.pagingUiListOf
 import com.stslex.wizard.core.ui.mvi.Store
 import com.stslex.wizard.feature.film_feed.ui.model.FilmModel
 import com.stslex.wizard.feature.film_feed.ui.model.ScreenState
-import com.stslex.wizard.feature.film_feed.ui.store.FeedStore.Action
-import com.stslex.wizard.feature.film_feed.ui.store.FeedStore.Event
-import com.stslex.wizard.feature.film_feed.ui.store.FeedStore.State
+import com.stslex.wizard.feature.film_feed.mvi.FeedStore.Action
+import com.stslex.wizard.feature.film_feed.mvi.FeedStore.Event
+import com.stslex.wizard.feature.film_feed.mvi.FeedStore.State
 
 interface FeedStore : Store<State, Action, Event> {
