+++ feature/film_feed/src/commonMain/kotlin/com/stslex/wizard/feature/film_feed/mvi/FeedStoreImpl.kt
@@ -1,14 +1,9 @@
-package com.stslex.wizard.feature.film_feed.ui.store
+package com.stslex.wizard.feature.film_feed.mvi
 
 import com.stslex.wizard.core.core.coroutine.AppDispatcher
 import com.stslex.wizard.core.ui.mvi.v2.BaseStore
 import com.stslex.wizard.feature.film_feed.di.FilmFeedScope
-import com.stslex.wizard.feature.film_feed.mvi.ClickersHandler
-import com.stslex.wizard.feature.film_feed.mvi.LoadFilmsHandler
 import com.stslex.wizard.feature.film_feed.navigation.FilmFeedComponent
-import com.stslex.wizard.feature.film_feed.ui.store.FeedStore.Action
-import com.stslex.wizard.feature.film_feed.ui.store.FeedStore.Event
-import com.stslex.wizard.feature.film_feed.ui.store.FeedStore.State
 import org.koin.android.annotation.KoinViewModel
 import org.koin.core.annotation.Qualifier
 import org.koin.core.annotation.Scope
@@ -23,18 +18,18 @@ internal class FeedStoreImpl(
     private val feedClickersHandler: ClickersHandler,
     private val loadFilmsHandler: LoadFilmsHandler,
     appDispatcher: AppDispatcher
-) : FeedHandlerStore, BaseStore<State, Action, Event, FeedHandlerStore>(
+) : FeedHandlerStore, BaseStore<FeedStore.State, FeedStore.Action, FeedStore.Event, FeedHandlerStore>(
     name = TAG,
-    initialState = State.INITIAL,
+    initialState = FeedStore.State.Companion.INITIAL,
     handlerCreator = { action ->
         when (action) {
-            is Action.Click.FilmClick -> feedClickersHandler
-            is Action.LoadFilms -> loadFilmsHandler
-            is Action.Navigation.Film -> component
+            is FeedStore.Action.Click.FilmClick -> feedClickersHandler
+            is FeedStore.Action.LoadFilms -> loadFilmsHandler
+            is FeedStore.Action.Navigation.Film -> component
         }
     },
     appDispatcher = appDispatcher,
-    initialActions = listOf(Action.LoadFilms),
+    initialActions = listOf(FeedStore.Action.LoadFilms),
 ) {
 
     internal companion object {
