+++ feature/film_feed/src/commonMain/kotlin/com/stslex/wizard/feature/film_feed/navigation/FilmFeedComponent.kt
@@ -4,8 +4,8 @@ import com.arkivanov.decompose.ComponentContext
 import com.stslex.wizard.core.navigation.Component
 import com.stslex.wizard.core.navigation.Config
 import com.stslex.wizard.core.ui.mvi.v2.Handler
-import com.stslex.wizard.feature.film_feed.ui.store.FeedHandlerStore
-import com.stslex.wizard.feature.film_feed.ui.store.FeedStore.Action.Navigation
+import com.stslex.wizard.feature.film_feed.mvi.FeedHandlerStore
+import com.stslex.wizard.feature.film_feed.mvi.FeedStore.Action.Navigation
 
 interface FilmFeedComponent : Component, Handler<Navigation, FeedHandlerStore> {
