+++ feature/film_feed/src/commonMain/kotlin/com/stslex/wizard/feature/film_feed/ui/FilmFeedScreen.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.Composable
 import com.stslex.wizard.core.ui.mvi.v2.NavComponentScreen
 import com.stslex.wizard.feature.film_feed.di.FilmFeedFeature
 import com.stslex.wizard.feature.film_feed.navigation.FilmFeedComponent
-import com.stslex.wizard.feature.film_feed.ui.store.FeedStore.Event
+import com.stslex.wizard.feature.film_feed.mvi.FeedStore.Event
 
 @Composable
 fun FilmFeedScreen(component: FilmFeedComponent) {
