+++ feature/follower/src/commonMain/kotlin/com/stslex/wizard/feature/follower/mvi/FollowerComponent.kt
@@ -1,4 +1,4 @@
-package com.stslex.wizard.feature.follower.navigation
+package com.stslex.wizard.feature.follower.mvi
 
 import com.arkivanov.decompose.ComponentContext
 import com.stslex.wizard.core.navigation.Component
