+++ feature/follower/src/commonMain/kotlin/com/stslex/wizard/feature/follower/ui/FollowerScreen.kt
@@ -7,9 +7,9 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
 import com.stslex.wizard.core.navigation.Config.Follower.FollowerType
 import com.stslex.wizard.core.ui.mvi.store_di.getStore
-import com.stslex.wizard.feature.follower.navigation.FollowerComponent
-import com.stslex.wizard.feature.follower.ui.store.FollowerStore
-import com.stslex.wizard.feature.follower.ui.store.FollowerStore.Action
+import com.stslex.wizard.feature.follower.mvi.FollowerComponent
+import com.stslex.wizard.feature.follower.mvi.FollowerStore
+import com.stslex.wizard.feature.follower.mvi.FollowerStore.Action
 import com.stslex.wizard.feature.follower.ui.store.FollowerStoreImpl
 import org.koin.core.parameter.parametersOf
